move entity to own project
