+++ Todo.Api.Tests/TodoApiTests.cs
@@ -1,3 +1,6 @@
+
+using TodoEntities;
+
 namespace TodoApi.Tests;
 
 public class TodoApiTests
