+++ Todo.Api/TodoDbContext.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
+using TodoEntities;
 
 namespace TodoApi;
