+++ Todo.Api/Todos/TodoApi.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.EntityFrameworkCore;
+using TodoEntities;
 
 namespace TodoApi;
 
@@ -30,7 +31,7 @@ internal static class TodoApi
         {
             return await db.Todos.FindAsync(id) switch
             {
-                Todo todo when todo.OwnerId == owner.Id || owner.IsAdmin => TypedResults.Ok(todo.AsTodoItem()),
+                { } todo when todo.OwnerId == owner.Id || owner.IsAdmin => TypedResults.Ok(todo.AsTodoItem()),
                 _ => TypedResults.NotFound()
             };
         });
