+++ Todo.Api/Todos/TodoItem.cs
@@ -1,0 +1,11 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace TodoApi;
+
+public class TodoItem
+{
+    public int Id { get; set; }
+    [Required]
+    public string Title { get; set; } = default!;
+    public bool IsComplete { get; set; }
+}
