+++ Todo.Api/Todos/TodoMappingExtensions.cs
@@ -1,0 +1,16 @@
+using TodoEntities;
+
+namespace TodoApi;
+
+public static class TodoMappingExtensions
+{
+    public static TodoItem AsTodoItem(this Todo todo)
+    {
+        return new()
+        {
+            Id = todo.Id,
+            Title = todo.Title,
+            IsComplete = todo.IsComplete,
+        };
+    }
+}
