+++ Todo.Entities/Todo.Entities.csproj
@@ -1,0 +1,11 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net9.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <AssemblyName>TodoEntities</AssemblyName>
+        <RootNamespace>TodoEntities</RootNamespace>
+    </PropertyGroup>
+
+</Project>
