+++ TodoApp.sln
@@ -31,6 +31,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{02EA681E-C7D
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{5C99C740-7E6A-4FFD-B21F-E807C980C205}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Todo.Entities", "Todo.Entities\Todo.Entities.csproj", "{0FF099DA-6A45-481B-9D27-E14E2E081CFE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -65,6 +67,10 @@ Global
 		{67B571B3-F4FB-46AA-888E-17A5A773610C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{67B571B3-F4FB-46AA-888E-17A5A773610C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{67B571B3-F4FB-46AA-888E-17A5A773610C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0FF099DA-6A45-481B-9D27-E14E2E081CFE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0FF099DA-6A45-481B-9D27-E14E2E081CFE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0FF099DA-6A45-481B-9D27-E14E2E081CFE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0FF099DA-6A45-481B-9D27-E14E2E081CFE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -77,6 +83,7 @@ Global
 		{272942F6-94E8-4D6B-8AD8-C4CCA305836D} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{4ECB456A-BC4A-4F99-A1D5-FE2DF1C4EA1D} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{67B571B3-F4FB-46AA-888E-17A5A773610C} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{0FF099DA-6A45-481B-9D27-E14E2E081CFE} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {28EAB679-0C71-494A-A1EC-D29DEDDD402E}
